[#38] refactor: Change modal location to organisms
