+++ src/components/index.ts
@@ -2,7 +2,7 @@ export { default as AppBootstrap } from "./app-bootstrap/app-bootstrap";
 export { default as Text } from "./atoms/text/text";
 export { default as ButtonNormal } from "./atoms/button/buttonNormal";
 export { default as ButtonGradient } from "./atoms/button/buttonGradient";
-export { default as Modal } from "./modal/modal";
+export { default as Modal } from "./organisms/modal/modal";
 // export { default as Form } from "./molecules/auth/form";
 // export { default as FormField } from "./molecules/auth/inputField";
 // export { default as SubmitButton } from "./molecules/auth/submitButton";
